+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 #### master
 
+- CHANGED: Updated GoDaddy parser to the new response (GH-60).
+
 - CHANGED: Updated Donuts parser to the new response. It looks like Donuts is now more compliant with base ICANN parser.
 
 - CHANGED: Updated Verisign parser to the new response (GH-57). [Thanks @phcyso]